requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-Cache==0.12
Jinja2==2.7.2
MarkupSafe==0.19
Werkzeug==0.9.4
argparse==1.1
beautifulsoup4==4.3.2
closure==20121212
colorama==0.2.7
cssmin==0.2.0
django-htmlmin==0.7.0
firapria==0.1.0
gunicorn==18.0
html5lib==0.999
itsdangerous==0.23
jsmin==2.0.9
lxml==3.3.3
newrelic==2.16.0.12
pep8==1.4.6
redis==2.9.1
six==1.6.1
webassets==0.9
wsgiref==0.1.2
xmltodict==0.8.6
